+++ commons/commons-distanceEstimation/src/main/java/com/lama/mse/commons/distanceEstimation/core/IDistanceEstimator.java
@@ -2,6 +2,6 @@ package com.lama.mse.commons.distanceEstimation.core;
 
 public interface IDistanceEstimator {
 
-	public int estimate(String startLocation, String endLocation);
+	public Long estimate(String startLocation, String endLocation);
 	
 }
